Move page serving plugin manager to visualization registry
